bot/requirements.txt
fastapi
uvicorn
openpyxl
python-dotenv
requests
pydantic
pandas
